Added Z_TILT_ADJUST
Added Z_TILT_ADJUST so the printer understands G34 just like Marlin.

G34 uses a Z probe to probe the edges of the bed and level the gantry

Uncomment the lines listed in the config if you wish to use Z_TILT_ADJUST and G34 but make sure to remove your Z sync belt before using.